:truck: Moved dev UI -> root UI
